fix Color::toArb

+++ src/gui/core/color.h
@@ -30,12 +30,12 @@ public:
     float alpha()const{return mComponents[3];}
     unsigned int toArgb()const;
 
-    static int red ( int c){return (c>>16)&0xFF;}
-    static int green(int c){return (c>> 8)&0xFF;}
-    static int blue (int c){return (c&0xFF);}
-    static int alpha(int c){return c>>24;}
+    static uint8_t red ( int c){return (c>>16)&0xFF;}
+    static uint8_t green(int c){return (c>> 8)&0xFF;}
+    static uint8_t blue (int c){return (c&0xFF);}
+    static uint8_t alpha(int c){return c>>24;}
     static unsigned int toArgb(float r, float g, float b, float a);
-    static unsigned int toArgb(int r,int g,int b,int a);
+    static unsigned int toArgb(uint8_t r,uint8_t g,uint8_t b,uint8_t a);
     static Color* valueOf(float r,float g,float b,float a=1.0);
     static unsigned int parseColor(const std::string&colorstring);
     static unsigned int getHtmlColor(const std::string&color);
